Move types and fix imports
