+++ app/layout.tsx
@@ -3,9 +3,8 @@ import './globals.css';
 import { Poppins } from 'next/font/google';
 
 import Header from '@/components/Header';
-
-import SupabaseProvider from '../providers/supabase-provider';
-import UserProvider from './user-provider';
+import SupabaseProvider from '@/providers/supabase-provider';
+import UserProvider from '@/providers/user-provider';
 
 const poppins = Poppins({
   weight: '400',
